src/requirements/dev.txt
-r base.txt
ipdb
mixer
django-debug-toolbar==1.6
